promemailserver/requirements.txt
djangorestframework
django
markdown
django-filter